Use default path until you don't
We start by following the combination of parent and slugified title, but if the user puts in something else for the path, that breaks the association and we won't override that. Unless the user sets it to match the parent and slugified title again, in which case we go back to tracking them.